Calculate Module Restructuring
Since coordinate based calcualtions currently (and likely in the future) play
a secondary role to PRISM based calculations, there doesn't seem to be a reason
to host those calculations under the same namespace as the primary calculations.
Coordinate based calculations which are added in the future will be hosted under
a util or coord module depending on need.