Removed Pre-Built Docs from Tracking
We don't want to have the built (html or pdf) documentation in our
version tracking because, among other reasons, it is redundant to the
files that are already being tracked. Furthermore, we could end up with
a mismatch between the version of the docs in the _build directory and
the repo.